build: Restructure how BuildCommand.build() is called
Rather than take a list of triplets to build, the BuildCommand.build()
function now takes a single repo/ref/morph triplet. Iterating through
multiple sets of triplets is now done in the build plugin.

There are a couple of cosmetic changes to the status output at the start
and end of a build as a result.