Allow `morph deploy` to cache missing Git repos if it needs to
This reverts commit c63f4e61f69820c71c2d8e9b96ce9bdec0d476a1.

This change was introduced back when the Baserock system definitions
used named Git refs rather than exact SHA1s. Morph would thus need to
update each repo involved in a build and resolve each ref before it
could build. Now that the system definitions use SHA1s, the 'updating
gits' phase is much less slow.

The problem with forcing 'no-git-update' is that it makes `morph deploy`
depend on the remote repo cache. In the case that all repos involved in
the build are avaiable in the remote repo cache (Trove), there is no
need for them to have been cached locally just to construct a build
graph. But if the network connection to the Trove is unreliable Morph
will need to fall back to the local repo cache. Unless Morph has done a
full local build of the system being built, it might not have every repo
cached locally.  Often, cached artifacts from the Trove or distbuild may
be used instead of locally building everything.

In this case the user sees `morph deploy` failing with errors such as:

    Repository upstream:binutils-redhat is not cached yet