Fix finding Git directories to include in a temporary build ref
When running 'morph build' the code looks through all the Git repos in
a system branch, to see which are involved in the build. These are then
checked for local changes and have temporary build refs created in them.

Due to a mistake in the logic, this would give up if it found a repo
that came from elsewhere but was inside the system branch directory. So
in the past some legitimate repos might have been ignored sometimes.