Fix `morph help-extensions` run outside a workspace
This command would use the sysbranchdir.open_from_within() function
call to search for a system branch, which traverses all subdirectories
looking for a directory. This is useful behaviour, but if the user ran
`morph help-extensions` in / it'd traverse their whole filesystem before
returning, which is stupid. This change means it only does the traverse
if it already detected a workspace.