Read files from a local clone of definitions where possible
Most morphologies involved in a build are in the definitions repo these
days. Currently we read each of them using `git cat-file`, which is
slow. It's quicker to check out all the files in one go to a temporary
directory and then read them from there.

With the current workflow users often have definitions.git checked out
on disk. It seems strange to not just read the files from there. There
are two reasons why I don't want to do that yet:
  - there are commands which don't run inside a system branch, which
    would be broken if we expected to always be in a system branch
  - there may be local changes in the checked-out repo, and it takes
    around 5 seconds on each build to check if there aren't any local
    changes. It actually seems faster to just check out a known clean
    version from the cache.