Treat path given to build cmd as relative to cwd
Currently the build commands treat the system argument
as a path relative to the root repo.

This means that regardless of your working directory you must run

morph build systems/foo-system.morph

This behaviour can be confusing, for example when your working
directory is $systembranch/definitions/systems you might expect
to be able to run

morph build foo-system.morph

especially since most shells would tab-complete the filename for you.

At the moment running the above command from $systembranch/definitions/systems
would result in an error, because morph would look for
$systembranch/definitions/foo-system.morph rather than
$systembranch/definitions/systems/foo-system.morph

This behaviour also means you can't give the morph build commands
an absolute path to a system morph.

This patch changes the treatment of the system arg so that it is interpreted
relative to the current working directory.