distbuild: Create a new directory to store build logs for each build.
It gets messy having hundreds of build-step-xx.log files in the current
directory, and if two builds are run in parallel from the same directory
the logs for a given chunk will be mixed together in one file.

Now, a new directory named build-0, build-1, build-2 etc is created for
each new build.

If the user passes --initiator-step-output-dir the logs will be placed
in that directory, instead. This behaviour is the same as before.