distbuild: Remove the build-steps message
This message was hundreds of kilobytes in size, as it contained a
recursive list of dependencies for each artifact in the build graph. It
was used in the initiator only to print this message:

    Build steps in total: 592

This message is now gone. The 'Need to build %d artifacts'
build-progress message now indicates the total build steps instead:

    Need to build 300 artifacts, of 592 total

This is a compatible change to the distbuild protocol: old initiators
will continue to work as normal with new controllers that don't send
the build-steps message.