sourceresolver: Fix LsTreeError raised when constructing build graph
If you had a repo that was not in your local or remote cache in your
definitions, Morph might raise LsTreeError and abort.

The code was assuming the repo would already be cached in the local
cache by the time it got to detecting build systems, but that's not
always true -- if the (reponame, ref) pair was already in the
'resolved_trees' cache but the repo had been deleted from the local
Git cache, to name one possibility.

Also, the remoterepocache.ls_tree operation can fail if the repo is
not hosted in the Trove, so we shouldn't abort the program in that case.

Finally, this patch removes an unused variable.