sourceresolver: Fix InvalidRefError when local git cache is out of date
Code in sourceresolver.py assumed that resolve_ref() would be called for
the chunk repo before get_morphology() was called, so the repo would
always be up to date. This wasn't actually true.

If your local repo cache was out of date, you might see the following
sort of error:

    InvalidRefError: Git directory
    /src/cache/gits/git___git_baserock_org_delta_usbutils has no commit at
    ref c37f146eb2c6642c600f1b025a6d56996b0697ff^{tree}.