Implement partial builds
This commit allows you to build only a single chunk or stratum and its
dependencies, instead of a whole system. This will be useful for
developers who want to quickly check the build commands of a chunk for
example. You can give more than one chunk or stratum to the command,
and the build will run up to the latest one, building the dependencies
of both.

An example of the usage:

    morph build systems/devel-system-x86_64-generic.morph \
        strata/build-essential.morph strata/core/curl.morph

This will only build up to curl in the core stratum.

Change-Id: Id55b8335fc61e8fdf847f5e383cfedd13d07e6d2