Enable native-bootstrap to continue build after recovered from fault.
The previous script creates new directory for the chunk it is going
to build without checking whether the directory exists or not.
It will fail back if the directory it attempted to creat exists.
So if build failed, you always need to remove all .inst directories and
let the native-bootstrap script build from the beginning.

This patch improves this, and you can run the native bootstrap script
direct after resumed without loss previous progress.

A condition was added to determine whether previous native bootstrap
script was terminated when it was building current chunk.
As .build directory for certain chunk only exists during building
phase of itself, it was created when started build, and cleaned up
after building finished. If .inst for certain chunk exists,
the .build directory doesn't, the building of the chunk should
be succeed in previous build.

The second go of native-bootstrap will skip all successful chunks
and start where it left off.

Change-Id: I91ae213ecc8c98808efdfd969624291e70f7e0fe