deploy: Deploy and upgrade systems from the same 'cluster' definition
This patch adds two fields to deployment (cluster) .morph files:
'upgrade-type' and 'upgrade-location'.

The `morph deploy` command ignores these. The `morph upgrade` command
will honour them if present, instead of the existing 'type' and
'location' fields. If they are not present, `morph upgrade` will give
a warning, and will use the 'type' and 'location' fields as before.

This avoids the need to edit the deployment .morph file after deploying
a system.

Small detail: the 'type' and 'location' variables are no longer removed
from the environment that is passed to the .configure and .write
extensions. This shouldn't affect anything.

Change-Id: Id2a4e4f229b8adebdb57eded2049ac113a82a4be