Make listing contents of local tarball cache more robust
Previously it would crash with a backtrace if there were unexpected
filenames in the directory. It's still not amazingly robust, but I
don't have time to rewrite the whole thing now.

This code seems to have ignored that cachefs.walkfiles() returns
filenames with a preceeding '/', which I have fixed now.

Change-Id: I98b3094bd6c82b26984513ee81a1eab9bf253a34