Add a CSV manifest generator.
This adds a `morph generate-manifest-csv` command which creates
a CSV manifest for a system containing license information,
upstream URL and a guess at the version of each component in a
system.

The existing `morph generate-manifest` is also renamed to
`morph generate-manifest-genivi` to improve the distinction
between the two commands which generate a manifest.

The basis of this commit was done by
Michael Drake <michael.drake@codethink.co.uk>.

Change-Id: I712de11eaa6e1564c5bae5f5f396e5e77484972f