Fix partial distbuilds of non-existant components
Currently, attempting to distbuild a component which is not in
the given system or doesn't exist at all will cause the full
system to be built, rather than an error raised. This is because
the logic which checks that all components were found is
completely nonsensical. This commit makes it actually check the
right thing.

Change-Id: Ide4d7e3fa5f71e433f3a7b7c8c387fe594c92e43