Ignore BuildProgress messages
Once building starts we close the json machine on the initiator,
but we may have received build progress events between processing
our build-started event and closing the json machine, since there
is not a nice way to tell the different types of build progress
apart (they all use BuildProgress) we will ignore all BuildProgress
messages for now.

A possible fix for this is to introduce GraphProgress messages so that
we can report the building of the graph without reporting other types
of BuildProgress ("Waiting for worker" or "Transferring artifact to cache")
that we're not interested in.

Note that we will still report build failures or build success, so if
there's a mistake in the definitions this will be reported before the
detach can occur, similarly if the system is already built this will
be reported before the detach happens.

Change-Id: Ia006ccfba826d2c91f4dea6c028ecdcb5a2b02d6