Improve 'Deploy failed as system is not yet build' error.
If a user had the wrong setting for artifact-cache-server then they
could build a system with 'morph distbuild', then be told it wasn't
built by 'morph deploy'. The logs do not show any info about which
cache server was used, so it looks like a Morph bug. This change should
make such a configuration problem much easier to spot.

Change-Id: I932ca3c5caa46990f9bcd77d75b3862f49e0dfb8