SourceResolver: Allow the resolution of multiple systems
The existing Source resolution code handles resolution of multiple
systems sufficiently.

It is not appropriate to then take this source pool and attempt to
create a build graph from it though, as the logical structure of the
input of what we want to build, and the logical structure of what we
will produce are conflated in the Source object.

If we do not intend to create a build graph from the Source Pool we
generate, then it is an appropriate data structure that may be used to
analyse changes in the input to a build.

Change-Id: If8e4a726f16f8aca000e59ecbbeb7d926cc08391