distbuild: Fix initiator hanging when protocol errors occur
If the initiator sends an invalid build-request message, it will now
exit with the following sort of error:

    ERROR: Failed to build baserock:baserock/definitions
    f2d78e9b7221bca65cba53af3f3b50d50d90628f
    systems/build-system-x86_64.morph: Invalid build-request message. Check
    you are using a supported version of Morph. This distbuild network uses
    protocol version 2.

Previously, the controller would log an error to its log file, but it
would not send any response to the initiator so the initiator would
hang forever.

Behaviour is the same as before for the case where the initiator sends a
build-request message with the wrong protocol version: the initiator
will exit with an error message.

Change-Id: I94fdee02bc701d4a679a0261b3c46dbdf14cfcaf