GitDir: Fix setting fetch url when push url is already on-disk
290483010cfc7945cd4483fadd1d98c3b83efb3 broke morph checkout, which
uses set_fetch_url on a repository that has been cloned, hence has its
origin remote url config already on-disk.

The fix prevents it changing the push_url when the fetch_url is set,
unless it is an unnamed remote, as if the config is on-disk, this
already does the right thing.

Change-Id: I6204f664407bab3d7f8ecf0fcca72f5015dee55e