Raise an error if a stratum build-depends on itself
If a stratum build-depends on itself, the build graph calculation
gets stuck in an infinite loop as it adds the same stratum to the
queue of morphologies to inspect over and over again.

This commit causes MorphologyLoader.validate_stratum to raise an
error if a stratum contains itself in it's build-depends, as
depending on itself makes no sense and will cause the above problem.

Change-Id: I76df5b7d63d010ae3b17f72bfa39b273e74279dd