Explain how to cancel a distbuild
Cancelling a distbuild with ctrl+c no longer cancels the build
itself. This commit adds some output explaining what should be
done to cancel the build as well as the local process.

This commit also fixes a bug where the BuildStarted event would
be sent each time a chunk finished building, since it was being
sent in _queue_worker_builds. This is fixed by adding a new
function to be called when the build graph annotation is
complete which sends BuildStarted and then calls
_queue_worker_builds, which no longer sends the BuildStarted
event.

Change-Id: I26ddea2c9080887f449e87004411ddffe4e583b7