generate-manifest-csv: Catch NoRemote error when cloning lorries repo
If there is no $trove-id/local-config/lorries repository, and similarly if
there is no baserock/local-config/lorries repository on the Trove defined
by 'trove-host', the generate-manifest-csv command will fail with a
NoRemote error from trying to clone a non-existent repository. This commit
fixes this and instead outputs a warning if either of the repositories
cannot be found.

Change-Id: I8edd8484e45d8fc91021a9be34966c13d2beb705