morph deploy: Allow overriding UPGRADE in definition or on command line
Concievably, you may want to force a cluster to behave as an upgrade
regardless of whether --upgrade is set or not.

This would allow two use cases that are not currently possible:

1.  To force a cluster to be upgrade-only, which would make our
    instructions for the upgrade-devel.morph cluster simpler.

2.  To allow a cluster to deploy then immediately upgrade.
    One potential use for this would be to create a disk image with
    multiple versions on it.

This also works as a work-around for --upgrade implying UPGRADE=yes for
subsystem deployments, but that should be fixed separately.

Change-Id: I2ec4b880fc9bce37adee9df67696f088c76650f4