Do not call str() on None types in sanitize_env
For a cluster morph,

- morph: systems/devel-system-x86_64-generic.morph
  deploy:
    self:
      type: ssh-rsync
      location: root@127.0.0.1
      ATTACK_KITTENS:

The yaml parser will assign the value None
to the element with key 'ATTACK_KITTENS',
sanitize_environment blindly calls str() on each value in the dictionary,
the result is that our extensions receive an 'ATTACK_KITTENS' environment
variable with the string value 'None', this is not useful and confuses anyone
trying to do basic validation in their extension.

This patch makes it so that any None types are replaced with the empty string,
which seems a more sensible value for this case.

Change-Id: I734d33e9304f0b7d4d590e464e350d2e5b4720d3