sourceresolver_tests: Remove bogus test
This test, tests for a case that produces unhelpful results.

It asserts that if we don't have the repository locally, and we try to
fetch it remotely and the file doesn't exist, that we should have no
morphology.

What we'd actually like is for it to fetch the repository and try that
locally, at which point it would succeed.

The reason the flawed previous behaviour (that this test asserted to be
correct) failed to cause problems for us, is that in the normal course
of resolving refs, we would always either have found it remotely or
cached it locally, when we were doing ls_tree or resolve_ref operations.

Change-Id: I287eb1e5bc9c916b31f306b4f5f24b0a6d3f0559