Extract downloaded tarball as current user
tar will preserve ownership and permissions by default
if running as the superuser, whereas git doesn't track ownership,
the owner is set to the user running the clone.

This is a reproducibility hole since the resulting repo may differ
depending on whether the repo was obtained from a tarball or via a git
clone, but morph considers the repos to be equivalent.

Running tar with --no-same-owner ensures the files
written by tar will be owned by the user extracting the archive.

This only fixes half the problem,
lorry may also need to be modified to ensure files in the
tarball have the same mode as those in the git repo.

Change-Id: I849d054bc4574f9c0bbcfd608914306e52b7a22e