Remove support for Baserock definitions format versions 0, 1 and 2
The previous release of the Baserock reference systems (baserock-15.25
tag) used Baserock definitions version 3, so this version of Morph can
still build the last release of the Baserock reference systems.

In the course of fixing up the cmdtest tests to use version 5
definitions, I removed a couple of bits of redundant code.

Change-Id: If9d480aa966d2ea5939cf04351f3d488cccea67f