Don't provide default strip-commands for 'manual' build system
This makes more sense to me conceptually: the idea of the 'manual' build
system is you specify *everything* manually.

However, I made this change to work around a problem in the Yarn test
suite. If the definitions used for testing are set to version 5 or
newer, `morph build` will try to run 'strip' and 'objcopy', which fails
because they don't exist in the test definitions.

We could solve this by setting 'strip-commands: []' in the test chunk
.morph files, but this triggers an error further down where `morph edit`
fails with 'Field strip-commands not allowed in morphology
test-chunk.morph'. This bug is caused by the conditional way that
strip-commands are currently implemented in Morph. When we drop support
for versions of the definitions format older than V5, the bug will go
away, and the `morph edit` command is deprecated in any case, so I think
it makes sense to work around it instead of trying to fix it.

Change-Id: Ib5124d72466a77cf3f28ed3e14cc4c231bdce4c4