Warn if system-integration commands are defined for an unknown artifact
The system-integration field in a chunk .morph file defines which of the
chunk artifacts the integration script should live. If you list an
artifact that doesn't exist, the integration script will not go anywhere
and the commands won't run.

Ideally system-integration commands will be become a bit more simple to
use, but hopefully this warning helps in the meantime.

Change-Id: I10a068e79eb46f98f0bd308a5caf3c445dda22ed