Simplify StagingArea class
Pass the Source to the staging area constructor so that we don't need to
pass it as a parameter when we call some StagingArea methods.  Also move
the creation of the build and destdir directories to the constructor so
that we can get rid of the chroot_open() and chroot_close() methods.

Also, provide API to retrieve the relative locations for buildir and
destdir.

Change-Id: I6e8085392e19ff3d8df807f260acf90eec9e0901