Improve get-repo error message
When attempting to get a repo with no commit at the ref we want,
we get a 'No such file or directory' error, which is confusing,
this is because we attempt to remove the directory we plan to clone
into, but if there's no commit at the ref we want then we don't
get as far as running git clone, so the directory doesn't exist,
and we fail to remove it.

This commit improves the error message by catching the InvalidRefError
separately.

Change-Id: Ibca20dd995d858713e56ab834daa97a8297cafc8