Don't require chunks in a stratum to appear before their dependencies
Currently Morph enforces that chunk A must be defined before anything
that build-depends on it. YBD doesn't enforce that.

Definitions format at <http://wiki.baserock.org/definitions/current/>
doesn't mention ordering currently. I propose that we make Morph be
permissive about ordering, like YBD is, and update the spec to mandate
no restrictions on ordering.

Since behaviour was previously undefined, making Morph be more
permissive about this shouldn't require a new version number of the
definitions format.

I still think we need to make sure stratum .morph files are ordered
logically, but that is in the realm of 'code style', it shouldn't be
being enforced by a build tool.

Change-Id: I425f2e5b9dfb62e4a26ed11f5c50e3978a0dd1a6