Use the ProgressBar class when downloading tarballs
The wget command already shows a progress bar but the text remains on
the tty after the download finishes.  By using ProgressBar, we show an
output that is consistent with the one when an artifact is downloaded,
and we can always show the progress bar as it will not pollute the tty
after the download finishes.

The progress bar is not shown when the --quiet option is given.

Also, change the units of the progress bar to be KiB and MiB.

Change-Id: I7e7685750f9619e76edb5700d51380e777a511fd