Cease modifying the morphologies
With `morph edit` removed, there is no need to load all the
morphologies, check whether a chunk was `morph edited`, change the
current ref to the build ref and write back the morphologies to a
temporary branch.

That is a lot of work as code profiling demonstrated.  With this patch
applied, morph execution finishes around 10 seconds sooner on my
machine.  This is not a big achievement when a full build is performed
where the wall clock time is dictated by the actual build commands, but
it will provide a much quicker feedback when the build artifacts were
already cached, or the semantic validation of morphologies fail.

We add the option `--untracked-files` to _get_status() in the GitIndex
class so that uncommitted morphologies on a uncommitted directory are
considered.  Previously this was being done by calling the following
call in inject_build_refs():

        self._root_index.add_files_from_index_info(
            self._hash_morphologies(self._root, morphs.morphologies))

This commit also removes some now unused code.

Change-Id: I14215db5c06ab06045ce901131e4e341271a039d