Cleanup buildbranch.py module
Now that we don't support system branches, we don't need to iterate over
a list of repos to create temporary build branches, commit their local
changes and push those branches.  We only need to do this for the
definitions repository itself.

A bug is also fixed where the local build branch was not being deleted
due a missing call to _register_cleanup() when commiting the changes.

This commit also renames some functions to more clear names, and moves
the logic on pushed_build_branch() to the definitions_repo module, the
only place where it is used.

Change-Id: Id86240d0c189245bed36bc46355be13d46498dbc